backend/requirements.txt
# =============================================================================
# Web Framework
# =============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# =============================================================================
# Environment & Configuration
# =============================================================================
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# =============================================================================
# HTTP Clients (for API calls to Bolna, Make.com, etc.)
# =============================================================================
httpx==0.25.2
requests==2.31.0

# =============================================================================
# AI & ML
# =============================================================================
# OpenAI (for LLM and embeddings)
openai==1.3.7

# Pinecone (vector database)
pinecone-client==2.2.4

# =============================================================================
# Database (Optional - for persistent storage)
# =============================================================================
# MongoDB
pymongo==4.6.0

# =============================================================================
# Utilities
# =============================================================================
python-dateutil==2.8.2
pytz==2023.3

# =============================================================================
# Optional: For advanced features
# =============================================================================
# Profanity filter
better-profanity==0.7.0

# For WhatsApp integration (via Twilio)
twilio==8.10.0

# For Google Sheets direct access (if not using Make.com)
# google-auth==2.23.4
# google-auth-oauthlib==1.1.0
# google-auth-httplib2==0.1.1
# google-api-python-client==2.108.0